feat: scaffold analytics sections
