+++ app/(app)/analytics/custom/page.tsx
@@ -1,0 +1,8 @@
+export default function CustomAnalytics() {
+  return (
+    <div className="p-6">
+      <h1 className="text-2xl font-semibold mb-4">Custom Analytics</h1>
+      <p>Saved analytics will be accessible here.</p>
+    </div>
+  );
+}
