+++ app/(app)/analytics/overview/page.tsx
@@ -1,0 +1,8 @@
+export default function AnalyticsOverview() {
+  return (
+    <div className="p-6">
+      <h1 className="text-2xl font-semibold mb-4">Analytics Overview</h1>
+      <p>Standardised visualisations will appear here.</p>
+    </div>
+  );
+}
